fix: move theme out into own file to reduce dep-cycle
DEVSU-1718
