+++ app/App.tsx
@@ -3,7 +3,7 @@
  */
 import { SnackbarProvider } from 'notistack';
 import CssBaseline from '@mui/material/CssBaseline';
-import { createTheme, ThemeProvider, StyledEngineProvider } from '@mui/material/styles';
+import { ThemeProvider, StyledEngineProvider } from '@mui/material/styles';
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 // eslint-disable-next-line import/no-extraneous-dependencies
@@ -14,7 +14,7 @@ import { CsvExportModule } from '@ag-grid-community/csv-export';
 import MainView from './views/MainView';
 import { SnackbarUtilsConfigurator } from './services/SnackbarUtils';
 import CacheBuster from './components/CacheBuster';
-import cssTheme from './styles/_theme.module.scss';
+import theme from './appTheme';
 
 // eslint-disable-next-line import/no-extraneous-dependencies
 import '@ag-grid-community/core/dist/styles/ag-grid.min.css';
@@ -22,75 +22,6 @@ import '@ag-grid-community/core/dist/styles/ag-grid.min.css';
 import '@ag-grid-community/core/dist/styles/ag-theme-material.min.css';
 import '@fontsource/roboto';
 import './styles/ag-grid.scss';
-
-const theme = createTheme({
-  direction: 'ltr',
-  palette: {
-    primary: {
-      main: cssTheme.primaryMain,
-      light: cssTheme.primaryLight,
-      dark: cssTheme.primaryDark,
-      contrastText: cssTheme.primaryContrastText,
-    },
-    secondary: {
-      main: cssTheme.secondaryMain,
-      light: cssTheme.secondaryLight,
-      dark: cssTheme.secondaryDark,
-      contrastText: cssTheme.secondaryContrastText,
-    },
-    error: {
-      main: cssTheme.errorMain,
-      light: cssTheme.errorLight,
-      dark: cssTheme.errorDark,
-      contrastText: cssTheme.errorContrastText,
-    },
-    text: {
-      primary: cssTheme.textPrimary,
-      secondary: cssTheme.textSecondary,
-      disabled: cssTheme.textDisabled,
-    },
-  },
-  typography: {
-    body1: { fontSize: cssTheme.fontSizeBody1 },
-    body2: { fontSize: cssTheme.fontSizeBody2 },
-    h1: { fontSize: cssTheme.fontSizeH1 },
-    h2: { fontSize: cssTheme.fontSizeH2 },
-    h3: { fontSize: cssTheme.fontSizeH3 },
-    h4: { fontSize: cssTheme.fontSizeH4 },
-    h5: { fontSize: cssTheme.fontSizeH5 },
-    h6: { fontSize: cssTheme.fontSizeH6 },
-    subtitle1: { fontSize: cssTheme.fontSizeSubtitle1 },
-  },
-  components: {
-    MuiCssBaseline: {
-      styleOverrides: {
-        '@global': {
-          html: {
-            WebkitFontSmoothing: 'auto',
-          },
-        },
-      },
-    },
-    MuiDialogTitle: {
-      styleOverrides: {
-        root: {
-          fontSize: cssTheme.fontSizeH2,
-        },
-      },
-    },
-    MuiButton: {
-      defaultProps: {
-        color: 'inherit',
-      },
-    },
-    MuiTabs: {
-      defaultProps: {
-        indicatorColor: 'secondary',
-        textColor: 'secondary',
-      },
-    },
-  },
-});
 
 ModuleRegistry.registerModules([
   ClientSideRowModelModule,
@@ -126,7 +57,3 @@ function App() {
 }
 
 export default App;
-
-export {
-  theme,
-};
