+++ app/appTheme.js
@@ -1,0 +1,73 @@
+import { createTheme } from '@mui/material/styles';
+import cssTheme from './styles/_theme.module.scss';
+
+const theme = createTheme({
+  direction: 'ltr',
+  palette: {
+    primary: {
+      main: cssTheme.primaryMain,
+      light: cssTheme.primaryLight,
+      dark: cssTheme.primaryDark,
+      contrastText: cssTheme.primaryContrastText,
+    },
+    secondary: {
+      main: cssTheme.secondaryMain,
+      light: cssTheme.secondaryLight,
+      dark: cssTheme.secondaryDark,
+      contrastText: cssTheme.secondaryContrastText,
+    },
+    error: {
+      main: cssTheme.errorMain,
+      light: cssTheme.errorLight,
+      dark: cssTheme.errorDark,
+      contrastText: cssTheme.errorContrastText,
+    },
+    text: {
+      primary: cssTheme.textPrimary,
+      secondary: cssTheme.textSecondary,
+      disabled: cssTheme.textDisabled,
+    },
+  },
+  typography: {
+    body1: { fontSize: cssTheme.fontSizeBody1 },
+    body2: { fontSize: cssTheme.fontSizeBody2 },
+    h1: { fontSize: cssTheme.fontSizeH1 },
+    h2: { fontSize: cssTheme.fontSizeH2 },
+    h3: { fontSize: cssTheme.fontSizeH3 },
+    h4: { fontSize: cssTheme.fontSizeH4 },
+    h5: { fontSize: cssTheme.fontSizeH5 },
+    h6: { fontSize: cssTheme.fontSizeH6 },
+    subtitle1: { fontSize: cssTheme.fontSizeSubtitle1 },
+  },
+  components: {
+    MuiCssBaseline: {
+      styleOverrides: {
+        '@global': {
+          html: {
+            WebkitFontSmoothing: 'auto',
+          },
+        },
+      },
+    },
+    MuiDialogTitle: {
+      styleOverrides: {
+        root: {
+          fontSize: cssTheme.fontSizeH2,
+        },
+      },
+    },
+    MuiButton: {
+      defaultProps: {
+        color: 'inherit',
+      },
+    },
+    MuiTabs: {
+      defaultProps: {
+        indicatorColor: 'secondary',
+        textColor: 'secondary',
+      },
+    },
+  },
+});
+
+export default theme;
